requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyascii==0.3.2
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
bangla==0.0.2
blinker==1.9.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dateparser==1.1.8
decorator==4.4.2
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.2.0
g2pkk==0.1.2
grpcio==1.67.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
hangul-romanize==0.1.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
inflect==7.4.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.8.4
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
networkx==2.8.8
nltk==3.9.1
num2words==0.5.13
numba==0.60.0
numpy==1.22.0
packaging==24.2
pandas==1.5.3
pathlib_abc==0.1.1
pathy==0.11.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
proglog==0.1.10
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pycparser==2.22
pydantic==1.10.13
pydub==0.25.1
pynndescent==0.5.13
pyparsing==3.2.0
pypdfium2==4.30.0
pypinyin==0.53.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
SudachiDict-core==20241021
SudachiPy==0.6.8
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thinc==8.1.12
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
tqdm==4.67.0
trainer==0.0.36
transformers==4.46.2
TTS==0.22.0
typeguard==4.4.1
typer==0.7.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
umap-learn==0.5.7
Unidecode==1.3.8
urllib3==2.2.3
wasabi==1.1.3
Werkzeug==3.1.3
yarl==1.17.1
